GWR conditions checked but still issues with representations
