+++ wrapify/config/manager.py
@@ -1,17 +1,6 @@
-import threading
 import yaml
 
-lock = threading.Lock()
-
-
-class SingletonOptimized(type):
-    _instances = {}
-    def __call__(cls, *args, **kwargs):
-        if cls not in cls._instances:
-            with lock:
-                if cls not in cls._instances:
-                    cls._instances[cls] = super(SingletonOptimized, cls).__call__(*args, **kwargs)
-        return cls._instances[cls]
+from wrapify.utils import SingletonOptimized
 
 
 class ConfigManager(metaclass=SingletonOptimized):
